Dodano pobieranie listu uzytkowników z filtrowaniem i stronicowaniem
poprawa zapytań


a

+++ src/main/java/pl/lodz/p/it/ssbd2022/ssbd02/mok/dto/ListDto.java
@@ -22,7 +22,7 @@ public class ListDto<T> {
     Integer recordsPerPage;
 
     @NotNull
-    Integer allRecords;
+    Long allRecords;
 
     @NotNull
     List<T> list;
